MAINT: restructured
